Separate chainer and pytorch
