+++ .gitignore
@@ -38,4 +38,6 @@ tools/warp-ctc*
 tools/sentencepiece/*
 tools/moses
 tools/*.done
+
 .idea/
+.hypothesis/
