+++ espnet/bin/lm_train.py
@@ -114,10 +114,10 @@ def main():
     # train
     logging.info('backend = ' + args.backend)
     if args.backend == "chainer":
-        from espnet.lm.lm_chainer import train
+        from lm.chainer.lm_chainer import train
         train(args)
     elif args.backend == "pytorch":
-        from espnet.lm.lm_pytorch import train
+        from lm.pytorch.lm_pytorch import train
         train(args)
     else:
         raise ValueError("Only chainer and pytorch are supported.")
